Move python package away from pyavd
