+++ .gitignore
@@ -6,11 +6,17 @@ __pycache__/
 .j2cache/
 *.log
 *.pickle
+.mypy_cache/
+
+# Python source
+/python/**/build/
+/python/**/dist/
+/python/**/*.egg-info/
+/python/**/.tox/
 
 # pyavd ignores
 /python-avd/build/
 /python-avd/dist/
-/python-avd/vendor/
 /python-avd/pyavd.egg-info/
 /python-avd/.tox/
 **/.coverage
@@ -59,12 +65,11 @@ python-avd/pyavd/*/schema/schema_fragments/.hash
 
 # Generated by Cargo
 # will have compiled files and executables
-/Cargo.lock
+Cargo.lock
 /target/
 /debug/
 /rust/**/debug/
 /rust/**/target/
-/rust/**/Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
