+++ .pre-commit-config.yaml
@@ -166,7 +166,6 @@ repos:
             .*\.svg|  # Exclude all .svg files
             .*\.excalidraw|  # Exclude all .excalidraw files
             .*\.rst|  # Exclude all .rst files
-            python-avd/pyavd/vendor/.*|  # Exclude all files in this directory
             ansible_collections/arista/avd/CHANGELOG.rst| # Exclude auto generated CHANGELOG.rst file
             ansible_collections/arista/avd/changelogs/.*|  # Exclude all files in this directory
             ansible_collections/arista/avd/roles/.*/docs/tables/.*\.md| # Exclude all .md files in tables
