+++ python/avdutils/MANIFEST.in
