+++ sonar-project.properties
@@ -4,6 +4,7 @@ sonar.organization=aristanetworks-1
 # Path to sources
 sonar.sources=\
     python-avd/,\
+    python/,\
     ansible_collections/arista/avd/plugins/
 
 # Exclude generated classes and ANTA related code while in preview
@@ -15,11 +16,13 @@ sonar.exclusions=\
     python-avd/pyavd/_cv/api/**,\
     python-avd/pyavd/_anta/**,\
     python-avd/pyavd/api/_anta/**,\
+    python/**/tests/**,\
     ansible_collections/arista/avd/plugins/action/anta_workflow.py
 
 # Path to tests
 sonar.tests=\
-    python-avd/tests/
+    python-avd/tests/,\
+    python/tests/
 
 # Python version
 sonar.python.version=\
@@ -30,4 +33,5 @@ sonar.python.version=\
 
 sonar.python.coverage.reportPaths=\
     python-avd/coverage.xml,\
+    python/avdutils/coverage.xml,\
     ansible_collections/arista/avd/*coverage.xml
